requirements.txt
aiohttp==3.10.10
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
distlib==0.3.8
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==64.0.0
django-cors-headers==4.4.0
django-filter==24.3
django-google-sso==6.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
isodate==0.7.2
itsdangerous==2.2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
python3-openid==3.2.0
referencing==0.35.1
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
rpds-py==0.20.0
urllib3==2.2.2
virtualenv==20.26.3
Werkzeug==3.0.4
win32-setctime==1.1.0
yarl==1.15.5
zeep==4.3.1
